refactor(jsonfacet): Changed access modifier to public for necessary classes and properties
