+++ solr/core/src/java/org/apache/solr/search/facet/FacetModule.java
@@ -328,23 +328,6 @@ public class FacetModule extends SearchComponent {
 }
 
 
-// TODO: perhaps factor out some sort of root/parent facet object that doesn't depend
-// on stuff like ResponseBuilder, but contains request parameters,
-// root filter lists (for filter exclusions), etc?
-class FacetComponentState {
-  ResponseBuilder rb;
-  Map<String,Object> facetCommands;
-  FacetRequest facetRequest;
-  boolean isShard;
-  Map<String,Object> facetInfo; // _facet_ param: contains out-of-band facet info, mainly for refinement requests
-
-  //
-  // Only used for distributed search
-  //
-  FacetMerger merger;
-  FacetMerger.Context mcontext;
-}
-
 // base class for facet functions that can be used in a sort
 abstract class FacetSortableMerger extends FacetMerger {
   public void prepareSort() {
